Move shell methods to a general helper
The helpers that interact with the user will need to be called from
outside of the `Cli` namespace in the future. I need classes loaded
before the `Cli` classes to be able to require a module like the `Shell`
one to access those methods.

I've added `module_function` to the module. This makes it possible to
call the methods on the module directly, without having to include the
module.

```
Shell.ask_for_input
Shell.required_input "Your name?"
Shell.yes_or_no "Cake?"
```
